Update for EXT:news & EXT:cb_newscal
Makes it possible to work with EXT:news & EXT:cb_newscal in conjunction.

Extends EXT:news with the field „ecom_evendate“ which will be used to
set a date for specific events (trade fairs, ecom events, …).
Additionally added the sorting ability for this field.

Now in EXT:cb_newscal you CAN use this field for setting the events in
the calendar.